requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.13.2
pydantic==2.11.7
openai
google-generativeai